resume_career_ai_app/requirements.txt
crewai
openai
streamlit
pymupdf
pdfplumber
requests
pandas
plotly
matplotlib
python-dotenv
